Audio+ some Math FFIs moved into ::FFI library.
