+++ 3DRadSpace/Engine3DRadSpace/Audio/AudioBuffer.cpp
@@ -1,5 +1,4 @@
 #include "AudioBuffer.hpp"
-#include "AudioBuffer.h"
 
 #include <al.h>
 #include <vorbis\codec.h>
@@ -248,40 +247,4 @@ std::expected<AudioBuffer, AudioBuffer::OGGLoadError> AudioBuffer::FromOGG(const
 	ov_clear(&vf);
 	fclose(file);
 	return AudioBuffer(buffer, channels, sampleRate, bps, format, static_cast<int>(size));
-}
-
-E3DRSP_AudioBuffer E3DRSP_AudioBuffer_Create(char* buffer, int numChannels, int sampleRate, int bps, int format, int size)
-{
-	assert(buffer != nullptr);
-	return new AudioBuffer(buffer, numChannels, sampleRate, bps, format, size);
-}
-
-E3DRSP_AudioBuffer E3DRSP_AudioBuffer_FromWAV(const char * path)
-{
-	assert(path != nullptr);
-
-	auto audioBuffer = AudioBuffer::FromWAV(path);
-	if(audioBuffer.has_value())
-	{
-		return new AudioBuffer(std::move(audioBuffer.value()));
-	}
-	else return nullptr;
-}
-
-E3DRSP_AudioBuffer E3DRSP_AudioBuffer_FromOGG(const char * path)
-{
-	assert(path != nullptr);
-
-	auto audioBuffer = AudioBuffer::FromOGG(path);
-	if(audioBuffer.has_value())
-	{
-		return new AudioBuffer(std::move(audioBuffer.value()));
-	}
-	else return nullptr;
-}
-
-void E3DRSP_AudioBuffer_Destroy(E3DRSP_AudioBuffer audioBuffer)
-{
-	assert(audioBuffer != nullptr);
-	delete static_cast<AudioBuffer*>(audioBuffer);
 }
