+++ 3DRadSpace/Engine3DRadSpace/FFI/Audio/IAudioEngine.cpp
@@ -1,5 +1,5 @@
-#include "IAudioEngine.hpp"
 #include "IAudioEngine.h"
+#include "../../Audio/IAudioEngine.hpp"
 
 using namespace Engine3DRadSpace;
 using namespace Engine3DRadSpace::Audio;
