+++ 3DRadSpace/Engine3DRadSpace/Objects/Impl/SoundEffect.cpp
@@ -171,58 +171,3 @@ REFL_FIELD_GS(SoundEffect, Looping, "Looping", false, "Is the sound looping")
 REFL_FIELD_GS(SoundEffect, Volume, "Volume", 1.0f, "Volume (0-1 range)")
 REFL_FIELD_GS(SoundEffect, Pitch, "Pitch", 0.5f, "Pitch (??-?? range), ?? default")
 REFL_END
-
-E3DRSP_SoundEffect E3DRSP_SoundEffect_Create()
-{
-	return new SoundEffect();
-}
-
-unsigned E3DRSP_SoundEffect_GetSoundID(E3DRSP_SoundEffect soundEffect)
-{
-	return static_cast<SoundEffect*>(soundEffect)->Sound.ID;
-}
-
-void E3DRSP_SoundEffect_SetSoundID(E3DRSP_SoundEffect soundEffect, unsigned soundAsset)
-{
-	static_cast<SoundEffect*>(soundEffect)->Sound.ID = soundAsset;
-}
-
-E3DRSP_Sound E3DRSP_SoundEffect_GetSound(E3DRSP_SoundEffect soundEffect)
-{
-	return static_cast<SoundEffect*>(soundEffect)->GetSound();
-}
-
-E3DRSP_SoundInstance E3DRSP_SoundEffect_GetSoundInstance(E3DRSP_SoundEffect soundEffect)
-{
-	return static_cast<SoundEffect*>(soundEffect)->GetInstance();
-}
-
-float E3DRSP_SoundEffect_GetVolume(E3DRSP_SoundEffect soundEffect)
-{
-	return static_cast<SoundEffect*>(soundEffect)->Volume;
-}
-
-void E3DRSP_SoundEffect_SetVolume(E3DRSP_SoundEffect soundEffect, float volume)
-{
-	static_cast<SoundEffect*>(soundEffect)->Volume = volume;
-}
-
-float E3DRSP_SoundEffect_GetPitch(E3DRSP_SoundEffect soundEffect)
-{
-	return static_cast<SoundEffect*>(soundEffect)->Pitch;
-}
-
-void E3DRSP_SoundEffect_SetPitch(E3DRSP_SoundEffect soundEffect, float pitch)
-{
-	static_cast<SoundEffect*>(soundEffect)->Pitch = pitch;
-}
-
-bool E3DRSP_SoundEffect_IsLooping(E3DRSP_SoundEffect soundEffect)
-{
-	return static_cast<SoundEffect*>(soundEffect)->Looping;
-}
-
-void E3DRSP_SoundEffect_SetLooping(E3DRSP_SoundEffect soundEffect, bool looping)
-{
-	static_cast<SoundEffect*>(soundEffect)->Looping = looping;
-}
